3738: create a common header files for AvmMini
